feat: split data managemnt page
